revert: restore CLI tools to 0.1.3 working state
- Remove network detection files and modules
- Restore simple tl-wrap.js architecture
- Remove test files and documentation for network detection
- Keep package.json dependencies for UI features
