+++ package.json
@@ -9,12 +9,9 @@
     "preview": "vite preview",
     "tauri": "tauri",
     "tauri:dev": "tauri dev",
-    "tauri:build": "tauri build",
-    "test": "vitest",
-    "test:ci": "vitest run"
+    "tauri:build": "tauri build"
   },
   "dependencies": {
-    "@anthropic-ai/claude-code": "^1.0.89",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -34,8 +31,7 @@
     "react-dom": "^19.0.0",
     "split2": "^4.2.0",
     "strip-ansi": "^7.1.0",
-    "tailwind-merge": "^3.3.1",
-    "vitest": "^3.2.4"
+    "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.1.12",
@@ -49,8 +45,7 @@
     "tailwindcss": "^4.1.12",
     "tw-animate-css": "^1.3.7",
     "typescript": "^5.7.2",
-    "vite": "^7.0.6",
-    "vitest": "^3.2.4"
+    "vite": "^7.0.6"
   },
   "engines": {
     "node": ">=20.19.0"
