--- tools/lib/ide-detector.js
@@ -1,62 +1,0 @@
-import { getIdeCommand, promptForIdeCommand } from './settings.js';
-import { execSync } from 'child_process';
-
-const IDE_MAPPINGS = {
-  'Visual Studio Code': 'code',
-  'Code': 'code',
-  'Cursor': 'cursor', 
-  'Windsurf': 'windsurf',
-  'WebStorm': 'webstorm',
-  'IntelliJ IDEA': 'idea',
-  'PyCharm': 'pycharm',
-  'PhpStorm': 'phpstorm',
-  'RubyMine': 'rubymine',
-  'CLion': 'clion',
-  'GoLand': 'goland',
-  'Rider': 'rider',
-  'Zed': 'zed',
-  'Xcode': 'xcode'
-};
-
-export function detectCurrentIDE() {
-  try {
-    if (process.env.VSCODE_INJECTION === '1' || process.env.TERM_PROGRAM === 'vscode') {
-      return getIdeCommand('Visual Studio Code', 'code');
-    }
-    if (process.env.CURSOR_AGENT || process.env.TERM_PROGRAM === 'cursor') {
-      return getIdeCommand('Cursor', 'cursor');
-    }
-    
-    const ppid = process.ppid;
-    if (ppid) {
-      try {
-        const parentName = execSync(`ps -p ${ppid} -o comm=`, { encoding: 'utf8' }).trim();
-        
-        const userCommand = getIdeCommand(parentName);
-        if (userCommand) {
-          return userCommand;
-        }
-        
-        if (IDE_MAPPINGS[parentName]) {
-          return getIdeCommand(parentName, IDE_MAPPINGS[parentName]);
-        }
-        
-        for (const [appName, command] of Object.entries(IDE_MAPPINGS)) {
-          if (parentName.toLowerCase().includes(appName.toLowerCase()) || 
-              appName.toLowerCase().includes(parentName.toLowerCase())) {
-            return getIdeCommand(parentName, command);
-          }
-        }
-        
-        const fallbackCommand = promptForIdeCommand(parentName);
-        return fallbackCommand;
-      } catch {
-      }
-    }
-  } catch {
-  }
-  
-  return null;
-}
-
-export { IDE_MAPPINGS };
